LayoutTests/svg/filters/feBlend-invalid-mode-expected.txt

CONSOLE MESSAGE: line 14: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: An invalid value was passed to an operation or assigned to an attribute.
PASS: Invalid blend modes do not trigger a crash. 
